Add files via upload
This is a software I created for a local company named "Loagma". I offered to develop a website and an admin panel that controls the website. The website is running but still requires some development. My favorite part of this project was to link the database with the admin panel and the website and make the content dynamic. I haven't charged the company as their vision is to help and support the local farmers of Telengana by bringing their product to the cities and making them a higher profit.